fix: fixes issues with iframe
