+++ package.json
@@ -23,6 +23,7 @@
     "@types/node": "^14.14.16",
     "@types/react": "^17.0.0",
     "@types/react-color": "^3.0.4",
+    "@types/react-frame-component": "^4.1.1",
     "@types/react-redux": "^7.1.14",
     "husky": "^4.3.6",
     "lerna": "^3.22.1",
