+++ packages/builder/components/views/sections/features/index.tsx
@@ -1,41 +1,25 @@
 import { FeaturesLayouts } from "@writy/sections";
 import React, { FC } from "react";
 import { useSelector } from "react-redux";
-import { getCurrentPageData, getSectionData } from "selectors/site";
+import { getSectionData } from "selectors/site";
 
 interface IProps {
   positionOfSection: number;
+  pageId: number;
 }
 
-const FeaturesSectionView: FC<IProps> = ({ positionOfSection }) => {
-  const currentPageId = useSelector(getCurrentPageData());
-  const sectionData = useSelector(
-    getSectionData(currentPageId, positionOfSection)
+const FeaturesSectionView: FC<IProps> = ({ positionOfSection, pageId }) => {
+  const { meta, data, theme } = useSelector(
+    getSectionData(pageId, positionOfSection)
   );
 
-  if (!sectionData) {
-    return null;
-  }
-
   const renderLayout = () => {
-    switch (sectionData.meta.layout) {
+    switch (meta.layout) {
       case "one":
-        return (
-          <FeaturesLayouts.One
-            theme={sectionData.theme}
-            data={sectionData.data}
-            meta={sectionData.meta}
-          />
-        );
+        return <FeaturesLayouts.One theme={theme} data={data} meta={meta} />;
 
       case "two":
-        return (
-          <FeaturesLayouts.Two
-            theme={sectionData.theme}
-            data={sectionData.data}
-            meta={sectionData.meta}
-          />
-        );
+        return <FeaturesLayouts.Two theme={theme} data={data} meta={meta} />;
 
       default:
         break;
